remove(hub-common): do not export private search utils
affects: @esri/hub-common

BREAKING CHANGE:
do not export private search utils
