+++ packages/common/src/search/_internal/portalSearchGroups.ts
@@ -11,7 +11,7 @@ import {
   IHubSearchResult,
   IQuery,
 } from "../types";
-import { getNextFunction } from "../utils";
+import { getNextFunction } from "./utils";
 import { expandPredicate } from "./expandPredicate";
 
 /**
