+++ packages/common/src/search/_internal/portalSearchItems.ts
@@ -16,7 +16,7 @@ import {
   IPredicate,
   IQuery,
 } from "../types";
-import { getNextFunction } from "../utils";
+import { getNextFunction } from "./utils";
 import { convertPortalAggregations } from "./portalSearchUtils";
 import { expandPredicate } from "./expandPredicate";
 import HubError from "../../HubError";
