+++ packages/common/test/search/_internal/expandApi.test.ts
@@ -1,0 +1,11 @@
+import { expandApi } from "../../../src/search/_internal/commonHelpers/expandApi";
+
+describe("expandApi", () => {
+  it("passes through objects", () => {
+    const chk = expandApi({
+      url: "https://my.enterprise.com/instance",
+      type: "arcgis",
+    });
+    expect(chk.type).toBe("arcgis");
+  });
+});
